resolve 64-bit Linux crash goodest rounding
Presumably, the issue has something to do with 64-bit Isaac's embedded
copy of Lua, how it handles rounding, and how the game expects these
values to be. Isaac on Linux is statically compiled, so this can't be
tested by replacing the Lua implementation with another (not easily, at
least)- this hack will have to do.